requirements1.txt
anaconda==0.0.1.1
certifi==2022.6.15
charset-normalizer==2.1.1
click==8.1.3
Flask==2.2.2
Flask-MySQL==1.5.2
idna==3.3
importlib-metadata==4.12.0
itsdangerous==2.1.2
Jinja2==3.1.2
MarkupSafe==2.1.1
mysql-client==0.0.1
PyJWT==2.4.0
PyMySQL==1.0.2
pytz==2022.2.1
requests==2.28.1
SQLAlchemy==1.4.40
twilio==7.13.0
urllib3==1.26.12
Werkzeug==2.2.2
zipp==3.8.1
